// src/utils/enum/status.enum.ts
export enum ClientStatus {
  POTENTIAL = 'POTENTIAL',
  FRECUENT = 'FRECUENT',
  UNSUBSCRIBED = 'UNSUBSCRIBED',
  COMPETENCE = 'COMPETENCE',
  COMPETENCE_FRECUENT = 'COMPETENCE_FRECUENT',
}

export enum RouteStatus {
  AVAILABLE = 'AVAILABLE',
  NEXT_VISIT = 'NEXT_VISIT',
  PENDING = 'PENDING',
  VISITED = 'VISITED',
}